Move dashboard to animals folder
